+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/KubernetesClient.java
@@ -155,4 +155,6 @@ public interface KubernetesClient extends Client {
 
   MixedOperation<LimitRange, LimitRangeList, DoneableLimitRange, Resource<LimitRange, DoneableLimitRange>> limitRanges();
 
+  SubjectAccessReviewDSL subjectAccessReviewAuth();
+
 }